% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Al2O3AnalysisApp-package.R
\docType{package}
\name{Al2O3AnalysisApp-package}
\alias{Al2O3AnalysisApp-package}
\alias{Al2O3AnalysisApp}
\title{Al2O3AnalysisApp: 'Shiny' Application for the Analysis of Al2O3:C Chips Measurements}
\description{
A 'shiny' application providing an exhanced graphical user interface to the function analyse_Al2O3C_Measurement() from the R package 'Luminescence'.
}
\details{
\tabular{ll}{
\strong{Version} \tab 0.1.2 \cr
\strong{License} \tab GPL-3 \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/Al2O3_AnalysisApp}
  \item Report bugs at \url{https://github.com/crp2a/Al2O3_AnalysisApp/issues}
}

}
\author{
\strong{Maintainer}: Archéosciences Bordeaux \email{services-archeosciences@u-bordeaux-montaigne.fr}

Authors:
\itemize{
  \item Sebastian Kreutzer \email{sebastian.kreutzer@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})
  \item Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})
}

Other contributors:
\itemize{
  \item Université Bordeaux Montaigne (\href{https://ror.org/03pbgwk21}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
  \item LabEx Sciences archéologiques de Bordeaux [funder]
}

}
\keyword{internal}
