% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_about_ui.R
\name{module_about_ui}
\alias{module_about_ui}
\title{About UI}
\usage{
module_about_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
About UI
}
\seealso{
Other modules:
\code{\link[=module_analysis_server]{module_analysis_server()}},
\code{\link[=module_analysis_ui]{module_analysis_ui()}},
\code{\link[=module_import_server]{module_import_server()}},
\code{\link[=module_import_ui]{module_import_ui()}},
\code{\link[=module_processing_server]{module_processing_server()}},
\code{\link[=module_processing_ui]{module_processing_ui()}},
\code{\link[=module_settings_server]{module_settings_server()}},
\code{\link[=module_settings_ui]{module_settings_ui()}}
}
\concept{modules}
\keyword{internal}
