% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis_ui.R
\name{module_analysis_ui}
\alias{module_analysis_ui}
\title{Analysis UI}
\usage{
module_analysis_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Analysis UI
}
\seealso{
\code{\link[=module_analysis_server]{module_analysis_server()}}

Other modules:
\code{\link[=module_about_ui]{module_about_ui()}},
\code{\link[=module_analysis_server]{module_analysis_server()}},
\code{\link[=module_import_server]{module_import_server()}},
\code{\link[=module_import_ui]{module_import_ui()}},
\code{\link[=module_processing_server]{module_processing_server()}},
\code{\link[=module_processing_ui]{module_processing_ui()}},
\code{\link[=module_settings_server]{module_settings_server()}},
\code{\link[=module_settings_ui]{module_settings_ui()}}
}
\concept{modules}
\keyword{internal}
