% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings_server.R
\name{module_settings_server}
\alias{module_settings_server}
\title{Settings Server}
\usage{
module_settings_server(id, user_settings)
}
\arguments{
\item{user_settings}{a \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Settings Server
}
\seealso{
\code{\link[=module_settings_ui]{module_settings_ui()}}

Other modules:
\code{\link[=module_about_ui]{module_about_ui()}},
\code{\link[=module_analysis_server]{module_analysis_server()}},
\code{\link[=module_analysis_ui]{module_analysis_ui()}},
\code{\link[=module_import_server]{module_import_server()}},
\code{\link[=module_import_ui]{module_import_ui()}},
\code{\link[=module_processing_server]{module_processing_server()}},
\code{\link[=module_processing_ui]{module_processing_ui()}},
\code{\link[=module_settings_ui]{module_settings_ui()}}
}
\concept{modules}
\keyword{internal}
