% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verification_hash}
\alias{verification_hash}
\title{Verification Hash}
\format{
A \code{\link{character}} vector.
}
\usage{
verification_hash
}
\description{
Codes are used to avoid that users upload weird sequences never approved and
then complain.
}
\note{
To create a new valid hash code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{temp <- Luminescence::read_XSYG2R(...)
digest::digest(names(temp))
}\if{html}{\out{</div>}}

If \code{temp} is list, then type \code{temp[[1]]}.
}
\concept{datasets}
\keyword{datasets}
\keyword{internal}
