% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Al2O3AnalysisApp-package.R
\docType{package}
\name{Al2O3AnalysisApp-package}
\alias{Al2O3AnalysisApp-package}
\alias{_PACKAGE}
\alias{Al2O3AnalysisApp}
\title{Al2O3AnalysisApp: 'Shiny' Application for the Analysis of Al2O3:C Chips Measurements}
\description{
A 'Shiny' application providing an exhanced graphical user interface to the function analyse_Al2O3C_Measurement() from the R package 'Luminescence'.
}
\details{
\tabular{ll}{
\strong{Package:} \tab Al2O3AnalysisApp \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. \href{https://lascarbx.labex.u-bordeaux.fr}{ANR-10-LABX-52}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/Al2O3_AnalysisApp}
  \item Report bugs at \url{https://github.com/crp2a/Al2O3_AnalysisApp/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Sebastian Kreutzer \tab \emph{Heidelberg University, Germany} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{services-archeosciences@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
