% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis_server.R
\name{module_analysis_server}
\alias{module_analysis_server}
\title{Analysis Server}
\usage{
module_analysis_server(id, user_data, user_settings)
}
\arguments{
\item{user_data}{a \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list with the
following elements: "\code{data}", "\code{info}" and "\code{samples}".}

\item{user_settings}{a \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Analysis Server
}
\seealso{
\code{\link[=module_analysis_ui]{module_analysis_ui()}}

Other modules: 
\code{\link{module_about_ui}()},
\code{\link{module_analysis_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_processing_server}()},
\code{\link{module_processing_ui}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{modules}
\keyword{internal}
