% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_processing_ui.R
\name{module_processing_ui}
\alias{module_processing_ui}
\title{Post-Processing UI}
\usage{
module_processing_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Post-Processing UI
}
\seealso{
\code{\link[=module_processing_server]{module_processing_server()}}

Other modules: 
\code{\link{module_about_ui}()},
\code{\link{module_analysis_server}()},
\code{\link{module_analysis_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_processing_server}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{modules}
\keyword{internal}
