% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings_ui.R
\name{module_settings_ui}
\alias{module_settings_ui}
\title{Settings UI}
\usage{
module_settings_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Settings UI
}
\seealso{
\code{\link[=module_settings_server]{module_settings_server()}}

Other modules: 
\code{\link{module_about_ui}()},
\code{\link{module_analysis_server}()},
\code{\link{module_analysis_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_processing_server}()},
\code{\link{module_processing_ui}()},
\code{\link{module_settings_server}()}
}
\concept{modules}
\keyword{internal}
