% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run a Shiny App}
\usage{
run_app(browser = TRUE, display = "auto")
}
\arguments{
\item{browser}{A \code{\link{logical}} scalar: should the app be run in
the browser?}

\item{display}{A \code{\link{character}} string specifying the mode in which
to display the application (see \code{\link[shiny:runApp]{shiny::runApp()}}).}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny:shinyApp]{shiny::shinyAppDir()}}.
}
\examples{
\dontrun{
run_app()
}
}
\author{
N. Frerebeau
}
\concept{shiny}
